+++ src/components/WatchFolder/mappingList/index.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import { WatchMapping } from 'types/watchFolder';
-import { MappingEntry } from './mappingEntry';
-import { MappingsContainer } from './styledComponents';
+import { MappingEntry } from '../mappingEntry';
+import { NoMappingsContent } from './noMappingsContent/noMappingsContent';
+import { MappingsContainer } from '../styledComponents';
 interface Iprops {
     mappings: WatchMapping[];
     handleRemoveWatchMapping: (value: WatchMapping) => void;
 }
 
 export function MappingList({ mappings, handleRemoveWatchMapping }: Iprops) {
-    return (
+    return mappings.length === 0 ? (
+        <NoMappingsContent />
+    ) : (
         <MappingsContainer>
             {mappings.map((mapping) => {
                 return (
