+++ src/components/WatchFolder/mappingList/noMappingsContent/checkmarkIcon.tsx
